move CPU packages to *-indef-cpu
