--- indef/core/src/Torch/Indef/TH/DynamicRandom.hs
@@ -1,7 +1,0 @@
-module Torch.Indef.TH.DynamicRandom (module X) where
-
-import Torch.Class.TH.Tensor.Random as X
-import Torch.Indef.TH.Dynamic.Tensor.Random as X
-
-import Torch.Class.TH.Tensor.Math.Random as X
-import Torch.Indef.TH.Dynamic.Tensor.Math.Random as X
