+++ indef/gpu/src/Torch/Indef/Static/Tensor/Random.hs
